fixup! ASoC: SOF: fix range checks
The mix of integer and size_t generates different results in 32- and
64 bit mode, cast to size_t and use %zu format.

+++ sound/soc/sof/control.c
@@ -232,8 +232,8 @@ int snd_sof_bytes_get(struct snd_kcontrol *kcontrol,
 	/* be->max has been verified to be >= sizeof(struct sof_abi_hdr) */
 	if (data->size > be->max - sizeof(*data)) {
 		dev_err_ratelimited(scomp->dev,
-				    "error: %u bytes of control data is invalid, max is %lu\n",
-				    data->size, be->max - sizeof(*data));
+				    "error: %u bytes of control data is invalid, max is %zu\n",
+				    data->size, (size_t)be->max - sizeof(*data));
 		return -EINVAL;
 	}
 
@@ -266,8 +266,8 @@ int snd_sof_bytes_put(struct snd_kcontrol *kcontrol,
 	/* be->max has been verified to be >= sizeof(struct sof_abi_hdr) */
 	if (data->size > be->max - sizeof(*data)) {
 		dev_err_ratelimited(scomp->dev,
-				    "error: data size too big %u bytes max is %lu\n",
-				    data->size, be->max - sizeof(*data));
+				    "error: data size too big %u bytes max is %zu\n",
+				    data->size, (size_t)be->max - sizeof(*data));
 		return -EINVAL;
 	}
 
@@ -441,8 +441,9 @@ int snd_sof_bytes_ext_get(struct snd_kcontrol *kcontrol,
 
 	/* check data size doesn't exceed max coming from topology */
 	if (cdata->data->size > be->max - sizeof(const struct sof_abi_hdr)) {
-		dev_err_ratelimited(scomp->dev, "error: user data size %d exceeds max size %lu.\n",
-				    cdata->data->size, be->max - sizeof(const struct sof_abi_hdr));
+		dev_err_ratelimited(scomp->dev, "error: user data size %d exceeds max size %zu.\n",
+				    cdata->data->size,
+				    (size_t)be->max - sizeof(const struct sof_abi_hdr));
 		return -EINVAL;
 	}
